requirements.txt
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
html2text==2025.4.15
idna==3.10
playwright==1.55.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
requests==2.32.5
rsa==4.9.1
soupsieve==2.8
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
